dailystonks-delivery/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.5
jinja2>=3.1
python-dotenv>=1.0
httpx>=0.27
email-validator>=2.1
python-multipart>=0.0.9
